SAA-2272: Replace alerts code
