+++ integration_tests/e2e/appointments/createRetrospectiveAppointment.cy.ts
@@ -14,7 +14,7 @@ import getScheduledEvents from '../../fixtures/activitiesApi/getScheduledEventsM
 import getAppointmentSeries from '../../fixtures/activitiesApi/getAppointmentSeries.json'
 import getGroupAppointmentSeriesDetails from '../../fixtures/activitiesApi/getGroupAppointmentSeriesDetails.json'
 import getGroupAppointmentDetails from '../../fixtures/activitiesApi/getGroupAppointmentDetails.json'
-import getOffenderAlerts from '../../fixtures/activitiesApi/getOffenderAlerts.json'
+import getPrisonerAlerts from '../../fixtures/alertsApi/getPrisonerAlerts.json'
 import HowToAddPrisonersPage from '../../pages/appointments/create-and-edit/howToAddPrisonersPage'
 import ReviewPrisonersPage from '../../pages/appointments/create-and-edit/reviewPrisonersPage'
 import ReviewPrisonerAlertsPage, {
@@ -73,7 +73,7 @@ context('Create a retrospective appointment', () => {
       '/users/jsmith',
       JSON.parse('{"name": "John Smith", "username": "jsmith", "authSource": "nomis"}'),
     )
-    cy.stubEndpoint('POST', '/api/bookings/offenderNo/MDI/alerts', getOffenderAlerts)
+    cy.stubEndpoint('POST', '/search/alerts/prison-numbers', getPrisonerAlerts)
     cy.stubEndpoint('POST', '/non-associations/between', [])
   })
