--- integration_tests/fixtures/activitiesApi/getOffenderAlerts.json
@@ -1,72 +1,0 @@
-[
-  {
-    "active": false,
-    "alertId": 1,
-    "bookingId": 830082,
-    "offenderNo": "A1351DZ",
-    "alertType": "X",
-    "alertTypeDescription": "Security",
-    "alertCode": "XNR",
-    "alertCodeDescription": "Not For Release",
-    "dateCreated": "2010-08-19",
-    "dateExpires": "2012-03-02",
-    "modifiedDateTime": "2017-05-09T21:50:35.311477",
-    "expired": true
-  },
-  {
-    "active": true,
-    "alertId": 2,
-    "bookingId": 830082,
-    "offenderNo": "A1351DZ",
-    "alertType": "X",
-    "alertTypeDescription": "Security",
-    "alertCode": "XTACT",
-    "alertCodeDescription": "Terrorism Act or Related Offence",
-    "dateCreated": "2023-03-20",
-    "dateExpires": "2012-03-02",
-    "modifiedDateTime": "2017-05-09T21:57:04.212529",
-    "expired": false
-  },
-  {
-    "active": true,
-    "alertId": 3,
-    "bookingId": 830082,
-    "offenderNo": "A1351DZ",
-    "alertType": "X",
-    "alertTypeDescription": "Security",
-    "alertCode": "XA",
-    "alertCodeDescription": "Arsonist",
-    "dateCreated": "2023-03-20",
-    "dateExpires": "2012-03-02",
-    "modifiedDateTime": "2023-03-20T21:57:04.212529",
-    "expired": false
-  },
-  {
-    "active": true,
-    "alertId": 4,
-    "bookingId": 830082,
-    "offenderNo": "A1351DZ",
-    "alertType": "R",
-    "alertTypeDescription": "Risk",
-    "alertCode": "RNO121",
-    "alertCodeDescription": "No 1 to 1 with this prisoner",
-    "dateCreated": "2023-03-20",
-    "dateExpires": "2027-03-20",
-    "modifiedDateTime": "2023-03-20T21:57:04.212529",
-    "expired": false
-  },
-  {
-    "active": true,
-    "alertId": 5,
-    "bookingId": 830082,
-    "offenderNo": "A1351DZ",
-    "alertType": "R",
-    "alertTypeDescription": "Risk",
-    "alertCode": "XCO",
-    "alertCodeDescription": "Corruptor",
-    "dateCreated": "2023-03-20",
-    "dateExpires": "2027-03-20",
-    "modifiedDateTime": "2023-03-20T21:57:04.212529",
-    "expired": false
-  }
-]
