--- integration_tests/fixtures/activitiesApi/getOffenderAlertsG0995GWG6123VU.json
@@ -1,16 +1,0 @@
-[
-  {
-    "active": true,
-    "alertId": 1,
-    "bookingId": 830082,
-    "offenderNo": "G0995GW",
-    "alertType": "X",
-    "alertTypeDescription": "Security",
-    "alertCode": "XNR",
-    "alertCodeDescription": "Not For Release",
-    "dateCreated": "2010-08-19",
-    "dateExpires": "2012-03-02",
-    "modifiedDateTime": "2017-05-09T21:50:35.311477",
-    "expired": false
-  }
-]
