+++ server/data/prisonApiClient.ts
@@ -2,7 +2,7 @@ import { Readable } from 'stream'
 import config, { ApiConfig } from '../config'
 
 import AbstractHmppsRestClient from './abstractHmppsRestClient'
-import { ReferenceCode, AgencyPrisonerPayProfile, Alert } from '../@types/prisonApiImport/types'
+import { ReferenceCode, AgencyPrisonerPayProfile } from '../@types/prisonApiImport/types'
 import { ServiceUser } from '../@types/express'
 import { LocationLenient } from '../@types/prisonApiImportCustom'
 
@@ -28,15 +28,6 @@ export default class PrisonApiClient extends AbstractHmppsRestClient {
   async getPayProfile(prisonCode: string): Promise<AgencyPrisonerPayProfile> {
     return this.get({
       path: `/api/agencies/${prisonCode}/pay-profile`,
-    })
-  }
-
-  // TODO: Move to prisonSearchApiClient once alerts endpoint is available
-  async getPrisonersAlerts(offenderNumbers: string[], prisonCode: string, user: ServiceUser): Promise<Alert[]> {
-    return this.post({
-      path: `/api/bookings/offenderNo/${prisonCode}/alerts`,
-      data: offenderNumbers,
-      authToken: user.token,
     })
   }
