+++ server/services/prisonService.ts
@@ -2,7 +2,7 @@ import _ from 'lodash'
 import { Readable } from 'stream'
 import PrisonApiClient from '../data/prisonApiClient'
 import PrisonerSearchApiClient from '../data/prisonerSearchApiClient'
-import { ReferenceCode, AgencyPrisonerPayProfile, Alert } from '../@types/prisonApiImport/types'
+import { ReferenceCode, AgencyPrisonerPayProfile } from '../@types/prisonApiImport/types'
 import { PagePrisoner, Prisoner } from '../@types/prisonerOffenderSearchImport/types'
 import { ServiceUser } from '../@types/express'
 
@@ -64,10 +64,6 @@ export default class PrisonService {
 
   async getPayProfile(prisonCode: string): Promise<AgencyPrisonerPayProfile> {
     return this.prisonApiClient.getPayProfile(prisonCode)
-  }
-
-  async getPrisonerAlerts(offenderNumbers: string[], prisonCode: string, user: ServiceUser): Promise<Alert[]> {
-    return this.prisonApiClient.getPrisonersAlerts(offenderNumbers, prisonCode, user)
   }
 
   async getPrisonerImage(prisonerNumber: string, user: ServiceUser): Promise<Readable> {
